+++ src/account.ts
@@ -1,101 +1,192 @@
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
-import { DynamoDBDocumentClient, GetCommand, PutCommand, DeleteCommand, QueryCommand } from "@aws-sdk/lib-dynamodb";
+import { DynamoDBDocument, UpdateCommandInput } from "@aws-sdk/lib-dynamodb";
 
 const client = new DynamoDBClient({ region: "us-west-2" });
-const dynamoDB = DynamoDBDocumentClient.from(client);
+const dynamoDB = DynamoDBDocument.from(client);
 
-const installationsKeyValueStore = 'Boost.GitHub-App.installations';
+const githubAppUserKeyValueStore = 'Boost.GitHub-App.installations';
+const reverseAccountLookupByUsernameSecondaryIndex = 'username-index';
 
-interface InstallationInfo {
+interface UserInfo {
     installationId?: string;
-    username?: string;
+    username: string;
     owner?: string;
     details?: string;
     lastUpdated?: number;
     authToken?: string;
 }
 
-export async function getUser(accountName: string): Promise<InstallationInfo | undefined> {
+export async function getAccountByUsername(username: string): Promise<string | undefined> {
     try {
         const params = {
-            TableName: installationsKeyValueStore,
+            TableName: githubAppUserKeyValueStore,
+            IndexName: reverseAccountLookupByUsernameSecondaryIndex,
+            KeyConditionExpression: 'username = :username',
+            ExpressionAttributeValues: {
+                ':username': username,
+            },
+        };
+
+        const item = await dynamoDB.query(params);
+        if (item.Items && item.Items.length > 0) {
+            // look for user info where the account is an email address
+            for (const user of item.Items) {
+                if (user.account.includes('@')) {
+                    return user.account;
+                }
+            }
+        }
+    } catch (error: any) {
+        console.error(`Error retrieving user info by username:`, error.stack || error);
+    }
+    return undefined;
+}
+
+export async function getUser(accountName: string): Promise<UserInfo | undefined> {
+    try {
+        const params = {
+            TableName: githubAppUserKeyValueStore,
             Key: {
                 account: accountName
             }
         };
 
-        const { Item } = await dynamoDB.send(new GetCommand(params));
-
-        if (Item) {
-            return {
-                installationId: Item.installationId,
-                username: Item.username,
-                owner: Item.owner,
-                details: Item.details,
-                lastUpdated: Item.lastUpdated,
-                authToken: Item.authToken
-            };
+        const item = await dynamoDB.get(params);
+        if (item.Item) {
+            return item.Item as UserInfo;
         }
-    } catch (error) {
-        console.error(`Error retrieving installation user info:`, error);
+    } catch (error: any) {
+        console.error(`Error retrieving user info:`, error.stack || error);
     }
     return undefined;
 }
 
 export async function saveUser(
     accountName: string,
-    installationId: string,
     username: string,
-    owner: string,
     installMessage: string,
-    authToken: string = ""): Promise<void> {
-    const params = {
-        TableName: installationsKeyValueStore,
-        Item: {
-            account: accountName,
-            installationId,
-            username,
-            owner,
-            details : installMessage,
-            lastUpdated: Math.round(Date.now() / 1000),
-            authToken
-        },
-    };
-    await dynamoDB.send(new PutCommand(params));
-}
-
-export async function deleteUser(username: string, isOrg: boolean): Promise<void> {
-    const queryParams = {
-        TableName: installationsKeyValueStore,
-        IndexName: 'username-index',
-        KeyConditionExpression: 'username = :username',
-        ExpressionAttributeValues: {
-            ':username': username
-        },
-        ProjectionExpression: 'account'
-    };
-
+    installationId?: string,
+    owner?: string,
+    authToken?: string): Promise<void> {
     try {
-        let accountName = username;
-        if (!isOrg) {
-            const { Items } = await dynamoDB.send(new QueryCommand(queryParams));
-            if (!Items || Items.length === 0) {
-                console.log(`No installation info found for username: ${username}`);
-                return;
-            }
-            accountName = Items[0].account;
-        }
-
-        const deleteParams = {
-            TableName: installationsKeyValueStore,
-            Key: {
-                account: accountName
-            }
+        // Build the update expression dynamically based on provided arguments
+        let updateExpression = "set lastUpdated = :lastUpdated, username = :username, details = :installMessage";
+        let expressionAttributeValues: any = {
+            ":lastUpdated": Math.round(Date.now() / 1000),
+            ":username": username,
+            ":installMessage": installMessage
         };
 
-        await dynamoDB.send(new DeleteCommand(deleteParams));
-        console.log(`Successfully deleted installation info for account: ${accountName}`);
-    } catch (error) {
-        console.error(`Error in processing deletion:`, error);
+        if (installationId) {
+            updateExpression += ", installationId = :installationId";
+            expressionAttributeValues[":installationId"] = installationId;
+        }
+        if (owner) {
+            updateExpression += ", owner = :owner";
+            expressionAttributeValues[":owner"] = owner;
+        }
+        if (authToken) {
+            updateExpression += ", authToken = :authToken";
+            expressionAttributeValues[":authToken"] = authToken;
+        }
+
+        const updateParams = {
+            TableName: githubAppUserKeyValueStore,
+            Key: { account: accountName },
+            UpdateExpression: updateExpression,
+            ExpressionAttributeValues: expressionAttributeValues,
+            ReturnValues: "UPDATED_NEW"
+        } as UpdateCommandInput;
+
+        await dynamoDB.update(updateParams);
+        console.log(`Successfully updated user info for account: ${accountName}`);
+    } catch (error: any) {
+        console.error(`Error saving user info for account: ${accountName}`, error.stack || error);
     }
 }
+
+export async function deleteUserByUsername(username: string, deleteInstallationInfoOnly: boolean = false): Promise<void> {
+    try {
+        // Query to find all accounts associated with the username
+        //      - This is necessary because the username is not the primary key
+        //      And there may be placeholder accounts missing the known primary email key
+        const queryResult = await dynamoDB.query({
+            TableName: githubAppUserKeyValueStore,
+            IndexName: reverseAccountLookupByUsernameSecondaryIndex,
+            KeyConditionExpression: 'username = :username',
+            ExpressionAttributeValues: {
+                ':username': username,
+            },
+        });
+
+        // If there are matching items, delete each one by its account name
+        if (!queryResult.Items || queryResult.Items.length === 0) {
+            console.log(`No user info found for username: ${username}`);
+            return;
+        }
+
+        for (const item of queryResult.Items) {
+            const accountName = item.account;
+            if (deleteInstallationInfoOnly) {
+                await updateUser(accountName, { username: username, installationId: "" });
+                console.log(`Successfully deleted installation info for account: ${accountName} for username: ${username}`);
+            } else {
+                try {
+                    await dynamoDB.delete({
+                        TableName: githubAppUserKeyValueStore,
+                        Key: { account: accountName },
+                    });
+                    console.log(`Successfully deleted user info for account: ${accountName} for username: ${username}`);
+                } catch (error: any) {
+                    console.error(`Error in deleting user info for account: ${accountName} for username: ${username}`, error.stack || error);
+                }
+            }
+        }
+    } catch (error: any) {
+        console.error(`Error in deleting user info for username: ${username}`, error.stack || error);
+    }
+}
+
+export async function deleteUser(accountName: string): Promise<void> {
+    try {
+        await dynamoDB.delete({
+            TableName: githubAppUserKeyValueStore,
+            Key: { account: accountName },
+        });
+        console.log(`Successfully deleted user info for account: ${accountName}`);
+    } catch (error: any) {
+        console.error(`Error in deleting user info for accountName: ${accountName}`, error.stack || error);
+    }
+}
+
+export async function updateUser(accountName: string, updatedInfo: UserInfo): Promise<void> {
+    let updateExpression = "set lastUpdated = :lastUpdated";
+    let expressionAttributeValues : any = {
+        ":lastUpdated": Math.round(Date.now() / 1000),
+    };
+
+    // Dynamically add fields to update based on what's present in updatedInfo
+    if ('authToken' in updatedInfo && updatedInfo.authToken !== undefined) {
+        updateExpression += ", authToken = :authToken";
+        expressionAttributeValues[":authToken"] = updatedInfo.authToken;
+    }
+    if ('installationId' in updatedInfo && updatedInfo.installationId !== undefined) {
+        updateExpression += ", installationId = :installationId";
+        expressionAttributeValues[":installationId"] = updatedInfo.installationId;
+    }
+
+    try {
+        const updateParams = {
+            TableName: githubAppUserKeyValueStore,
+            Key: { account: accountName },
+            UpdateExpression: updateExpression,
+            ExpressionAttributeValues: expressionAttributeValues
+        };
+
+        await dynamoDB.update(updateParams);
+        console.log(`Successfully updated user info for account: ${accountName} - ${JSON.stringify(updatedInfo)}`);
+    } catch (error: any) {
+        console.error(`Error in updating user info for account: ${accountName} - ${JSON.stringify(updatedInfo)}`, error.stack || error);
+    }
+}
+